To many elements to index (#13868)
* To many elements to index

I don't know why in the query `#__tags` is twice. Once as a `a` and second time as a `b`. For me it works perfectly when we remove `b`. I had a problem with this - for 1100 tags on my websites, Smart Search Indexer found more than 1233000 items to index.

* Mistake with query

I corrected a bug with calling the `#__tags` table in the query.