Greek stemmer fix
Rebasing to the LGPLv2.1 Drupal stemmer.

So, here's the funny thing. The only part of the code which I kept is,
in fact, identical to the LGPLv2.1 version of the code. The only parts
which were different between the two versions (LGPLv2.1 and GPLv3) of
the Drupal stemmer are irrelevant since they BOTH require ISO-8859-7
encoding which is incompatible with Joomla! anyway. I have written my
own case converter for the stemmer which is of course licensed under
whatever I want. Since I've signed the Joomla! Contributor Agreement
all those years ago I am licensing it "GPLv2 or later" and
contributing it to Joomla!. Licensing problem solved!