Sample Data not available message
If no sample data plugins are enabled the message displayed was very generic. This PR adds a more informative message. Hopefully this message will also be usable if/when the plugins are updated to not be displayed if they can not be installed eg if there is already content on the site