[4.0] [a11y] Mod-popular, mod_logged, mod_latest (#20962)
* [4.0] [a11y] Mod-popular - proof of concept

The modules on the control panel contain multi-dimensional data and so should be tables and not lists. By doing this we are also able to make them accessible.

This PR is a proof of concept and the actual markup (the classes) will neeed to be reviewed before merging by someone other than me.

### Accessibility improvements
Each column has a table header and each table header has a scope
The cell for the article name has a scope of row (it therefore has to be a th not a td)
Adding the scope ensures that screenreaders understand what the numbers are eg 1024 hits for article link "similar tags"
If the site owner choses not to display the module title then a sr-only caption is created

For completeness below are two screenshots. One showing all the options applied and the other if there are no articles

* table id

* cleanup - remember this is only a POC

* ensure the table id is unique

* strings

* really should be date published

* publish date not created date

* mod_latest

* oops forgot tbody

* mod_logged

* role button

* cs - thanks @quy

* cs

* fix link