[4.0][a11y] Content Navigation plugin (#21144)
* [4.0][a11y] Content Navigation plugin

This pr adds the role navigation to the next and previous buttons when the content navigation plugin is enabled. It also ensures that the article title is announced to the screenreader. I added a class to the nav to assist with future styling although its not got any right now.

When testing ensure it works correctly both when the plugin is set to use titles and when it is using next/prev

* dont need role=navigation with html5