[4.0] [a11y] Toggle Editor button (#21195)
This is a clear example of something that should be a button and not an "a" styled as a button.

In some cases it is ok to add a role=button to an "a" but there is one issue with that technique - an "a" only responds to an enter key. a button responds to an enter and a space. So if you use an "a" you have to add something to monitor the space.

In this specific case it is super easy to change from an "a" to a "button"

To test make sure the toggle editor button still works. You can also test if it can be activated by tabbing to it and then selecting with either the space or the enter keys.