[4.0] [a11y] control panel table modules (#21197)
In #20962 I changed the latest, logged, and popular modules to tables to improve a11y

I did add a caption for the table but only for when the module heading was not displayed - this was wrong

When using a screen reader such as nvda you can navigate between tables by pressing the T key. When you do this the heading is not announced as it is before the table - so the user does not know what the purpose of the table is.

This PR makes the table caption always available so that when navigating with the T key the user will know the purpose of the table.