[4.0] Remove css rule
Since the change to use custom elements for toolbar buttons we are still left with some toolbar buttons that are not using custom elements eg the batch button

All buttons in the toolbar (custom elements and old style) have a margin-left of 0.75rem

Except for the first occurrence of the old style button which had a margin-left of 0. As that first occurrence now no longer occurs at the beginning of the row it results in two buttons without a margin between them. (see image)

This quick pr deletes that first occurrence rule so that we have the space again (see image)

Any comment about why the batch is not a custom element is beyond the scope of this pr