Update en-GB.com_config.ini
Saying Server time zone gives a very wrong impression. The timezone to be set here has to be the Timezone for the Primary Audience of the website. Saying Server can confuse the end users. I might have bought hosting in the US so my Server is there.. But my website's primary Audience is in India. The current nomenclature can confuse people into setting the wrong timezone