Improved performance and ordering
Admin control panel very slow on system with a large number of guests. guest column is not indexed, added userid to where clause in order to speed up query by using an index. Also without an order by the order of rows returned is not guaranteed, so adding an order by ensures the latest X rows are returned.