[4.0] POC html5pattern validation
With J4 we have the ability to use html5pattern validation. Where relevant this is much better than using php validaton. php validation only takes place on submit so
- you have to wait until the end to know you made a mistake
- the resulting error message is not inline with the field that is wrong - very important for a11y and ux

[more info about html5patterns here #18995]

This POC adds the ability to use JText for the validationmessage

For purely test purposes I have updated one field - the username for the Super User account in the installer.

If this concept is accepted then it can be applied in many places throughout joomla

Note we have different validation routines currently (js and php) and they are not consistent nor do they always work as intended see report by @astridx #19301