[4.0] Workflow - list and edit
I can't see any point in having the author or creation/modified dates. We dont have them for stages/transitions and unless I have missed something they are not needed.

So this PR simplifies the edit form (inline with #21952) and updates the list view accordingly

NOTE I removed the manage text link as its really a duplicate of the other stages link that has the number in it - although that might need tweaking in the UI - as you should not have two links to do the same thing for a11y