[4.0] [a11y] Hide empty nav
This pr moves the if statement outside of the NAV container so that we dont end up with an empty nav when there is no pagination

To test go to any list view in the admin which does not have an enough entries for their to be any pagination. View source and you will find a `<nav>`container with no content. After the PR the container is not present.