[4.0] Table Column alignment (part 1) [a11y]
com_associations
com_banners
com_cache
com_contacts
com_content
com_contenthistory

The alignment of columns has changed since J3 and also doesn't match the design for j4. In addition text-center alignment is not good for accessibility, particularly dyslexi8a (word-blindness).

https://www.w3.org/WAI/tutorials/tables/tips/
Alignment: Align text to the left and numeric data to the right (in left-to-right languages), so that people using larger text sizes or smaller screens will be able to find it. It’s helpful to give column headers the same alignment as the data in the cells below.