Create "cache" folder if it doesn't exist before trying to write to it
This fixes this issue:

### Steps to reproduce the issue

- Have cache enabled
- Delete cache folder

### Expected result

- Joomla creates "cache" folder if it doesn't exist

### Actual result

- Sites breaks. Goes blank.

### Additional comments

Now, of course there is no reason why one should delete the cache folder but it can happen by accident or for whatever other reason. The principle that Joomla doesn't check if the folder is there or not (and breaks the site if it's not) before trying to write to it is wrong.