[4.0] [a11y] Rules permissions - text filters notes (#21559)
* [4.0] Rules [a11y]

I am not sure what the logic/reasoning was for putting the notes on the permissions tab inside a custom element "joomla-alert" but it was wrong - it is not an alert it is just information. Making it an alert has significant impact.

This custom element adds role="alert" to the container. In a11y this role=alert is equivalent to using aria-live=assertive. aria-live="assertive" should be used only when you think the user needs to know the information the very second it appears.

So what happens is that as soon as the permissions tab gets focus and before anything else can happen the entire content of the alert is announced (and cannot be interrupted because it is so important).

This PR changes it back to a simple div and adds css to match the current styling - which obviouslly can be changed and improved upon.

* the sccs

* hound

* put it back

* add text-filters