[4.0] Correct spacer in template options
As can be seen in the screenshot the "useless" title is in the wrong place. Looking further this field was completely wrong and even had a default value that is never used.

This pr changes the title to a spacer line - this is more consistent with the rest of joomla and removes that unused default value and the now unused language strings