[4.0] configuration options
If you go to the options of an article then the fields are rendered by `$field->renderfield`
BUt if you go to the options of the component then each field is rendered by `$field->label` and `$field-
>input`

This pr changes that so the component options are also using `$field->renderfield` which they should be.

### testing
Check that showon etc still works

### notes
This pr will expose some descriptions previously only available in a tooltip. The tooltip is removed by #22165 and a future pr will look at potentially removing some of those descriptions if they are no longer needed