This is a combination of 10 commits. This is the 1st commit message:
First commit which integrates the subform custom field type.

This is the commit message #2:

Rename manifest file for plg_fields_subform

This is the commit message #3:

plg_fields_subform: Improve code comments and made the subform field type optionally non-repeatable.

This is the commit message #4:

Fix not being able to change the type of a subfield.

This is the commit message #5:

Implement a small in-memory renderCache for subform custom fields

This is the commit message #6:

Prepare to load the subfields into a subform custom field directly via their own xml definition.

This is the commit message #7:

After merge with PR #17552 we can now successfully show the subforms options for the different types.

This is the commit message #8:

The subform custom field type doesnt need a default value.

This is the commit message #9:

Implement an option for a subform customfield to disable rendering the field values.

This is the commit message #10:

Remove unused code and set hidden fields to not-required.

Signed-off-by: Rene Pasing <r.pasing@continga.de>