Manual selection of tags in module Tags Popular
In mod_tags_popular.xml I added an extra option to the parameter field order_value. If this option is selected, the module shows the tags in the order as set manually in the Tags component. This can be combined with setting the maximum number of tags in order to create your custom selection (up to the overall maximum of 20). The idea for this functionality comes from a request from a site owner (one of my customers).

Please note that in this case the module does not take into account the "popularity" of the tag. This is exactly as I intended but it may conflict with the idea of the module: showing popular tags. However, the Random option of the module doesn't take that into account either (as a matter of fact, when selecting my cusom option the execution of the module script follows the same 'logical path' through the code as when selecting the Random option).

It might be a good idea to add some instruction about this. Apart from that, I think this offers some nice extra fuctionality and I was surprised at how easy it was to add this.