[4.0] Editor buttons [a11y]
this Pr is for the joomla xtd-editor buttons that appear BELOW the editor when not using tinymce

Before this pr they were `<a href= .... role=button`

This is fine but they are buttons, they are visually styled as buttons and they are identified as buttons to assistive technology.

However the default behaviour for a button is for it to be selected by either space or enter whereas the default behaviour for a link is for it only to be selected by the enter key.

This could be fixed by jvascript but far better to use the correct html and make the button really a button

There is no visual change - however if you use the tab key to navigate to the buttons you will be able to test that you can select the button and a modal opens by pressing the space bar as well as the enter key.