[4.0] joomla-field-media [a11y]
I have no idea how this passed an a11y review - it was not possible to use the keyboard to navigate the field to either the select or clear buttons AND the clear button had no textual description. In addition although they are buttons they have been created as links so that means they are not accessible by the space AND enter keys as a button should be.

AFAICT the buttons are now correct and the modal for selecting images opens etc but I dont understand the js in joomla-field-media.js to be 100% certain there are no unforeseen consequences.

Note the media manager itself _is completely unaccessible_