[4.0] btn-secondary missing css (#22615)
The css for this btn-secondary class was missing. If you go to administrator/index.php?option=com_menus&view=menus and click on a module button then the text disappears when you hover on it as its white text on a transparent baackground. This fixes that

Cannot be tested just with patchtester