Update index.php to solve bug #23148
`__DIR__` constant can hold `/` if the PHP file in the root directory. 
`__DIR__`  The directory of the file. If used inside an include, the directory of the included file is returned. This is equivalent to dirname(__FILE__). This directory name does not have a trailing slash unless it is the root directory.