Custom Fields - display units
It is very common to need to display a unit of measurement with a field. Currently there is no way to do this without creating custom layouts for every field and every unit when needed.

This PR adds a new param to every custom field "Suffix" which allows you to add a unit of measurement (or any other text) to be rendered after the value of the field.

The new field is disabled by default so three is no b/c issue

### To test
Apply the pr and then create a new custom field and you will see the new Suffix param in the render options.

## Example usage
### Before

### After