Allow the fielname to be overriden when using a subform form field.
The fieldname of a subform is being used to construct the prefix of the array keys for
the subform rows. See e.g. layout joomla.form.field.subform.repeatable-table.section
where $basegroup is being passed as data-base-name, which comes from the layout
joomla.form.field.subform.repeatable-table, where $fieldname is being passed as value.
This $fieldname is coming $displayData, which comes from JFormFieldSubform::getInput(),
which takes $this->fieldname for $data['fieldname']. So we need to be able to override
the JFormField::fieldname attribute on a subform, which by default is not possible,
see also JFormField::setup() where it is not possible to set the fieldname. So we do this
in the subform type directly.