Explicitly state our argument separator when calling http_build_query().
When the PHP arg_separator.input configuration option is not set to &
(e.g. if it's set to &amp;) then this will result in malformed encoded
strings. Sometimes this is generally OK, but when using
http_build_query() to format POST data (e.g. in the Http/Transport
classes) this can result in malformed POSTs and failing APIs

This is easily seen in e.g. reCaptcha verification requests.