[4.0] Disable status module links
In Joomla 3.x we disabled the user links in the top right hand corner of the admin at the same time that we disabled the main menu. Amongst other reasons this was to prevent people from leaving an item checked out.

In Joomla 4 we have now have several icons in this place (users, private-messages, post-installation, multilingual status, preview) and these are from the status module and the multilingual status module.

Nothing needs to be changed with the multilingual status module as this opens in a model and preview is also ok because it opens a new tab.

But users, private-messages, post-installation all need to be disabled when an item is open to prevent the item being left checked out.

In J3 and in J4 we have code for this ```$app->input->get('hidemainmenu');```

In J4 the variable is declared but never used - whoops :)

This PR fixes that.

### Test
Open an article and ensure that only the "preview" and if enabled "multilingual status" links work
Close the article and ensure that all of them work.

### Note
The template design team might want to change the icon or styling when it is disabled but that's beyond the scope of this PR.