[4.0] Remove unused string (#23750)
I assume that this string was intended to be used for the dropdown menu that is in the top right hand corner. IIRC it was at one point in atum's life hard coded into the template. However today that user menu is controlled by the mod_status administrator module and the language string used is from there and not from the template