[4.0] Disable Back to Control Panel link
In Joomla 3.x we disabled the control panel link in the top left hand corner of the admin at the same time that we disabled the main menu. Amongst other reasons this was to prevent people from leaving an item checked out.

In Joomla 4 the link is always active :(

This PR fixes that.

### Test
Open an article and ensure that the control panel link in the top left hand corner of the admin does not work
Close the article and ensure that the link works

### Related PR
See #23752

### Note
The template design team might want to change the icon or styling when it is disabled but that's beyond the scope of this PR.