[4.0] Remove description label
PR for #23874

When the only content of a tab is an editor there is no need to have the label displayed next to the editor instead it can be displayed above the editor. This is already the case for categories. This PR adds the same for tags and contacts (miscellaneous information)