Article Module comments
If we are going to have comments in code then they should be correct. This PR changes the position of a comment so that it is clear which options it refers to. I wasted hours debugging the code looking for a module param that the comment said should exist and it didnt becauser the comment was wrong