Update default.php
For A11Y reasons, as reported by Stefan Wadja and because in Web browser situations, like this, a 'Cancel' button is not needed, as the user can simply navigate away, if they don't want to commit, the 'Cancel' button is removed. In any case, there is no 'cancel' type for the <button> element, in HTML, so this can only be done by using CSS to make a <a> element look like a button 0r by making a normal button and adding Javascript. Neither is a good solution for an unnecessary button.

Therefore, it is being removed.