[4.0] Remove article print and email
The option to email a link to a friend and to print an article were outdated when joomla 1.0 was launched - there are far better ways to share content etc.

This pr removes the show print and email icons options and code to display the icons. (It does not remove the helpers as they haven't been deprecated)

If I understand everything correctly nothing would need to be done to update upgrading sites databases as the stored param would simply be ignored

Resurrection of PR#19393