[4.0] Remove com_admin profile
The users icon in the mod_status module links to a view of the logged in users profile. However because it is NOT using com_users any plugins that extend com_users are not displayed. For example TFA plugins when enabled are not accessible from this link and are only accessible from the users component. Maintaining this additional view makes no sense to me. Its a regular cause of bug reports etc.

This PR completely removes the com_admin profile and replaces the link in the module to use com_users. Now any plugins that extend com_users are accessible from the link which is what the user expects