[4.0] Change Status dropdown consistency
Having a consistent UI is essential. All off the new change status dropdown items are ordrered Published->Unpublished->Archive->Checkin->Trashed except for com_content where Archive and Checkin are reversed.

This PR changes the order of the articles change status dropdown items to be consistent with all the others