[4.0][a11y] RFC Adaptive Accessibility Menu
This is a proof of concept for evaluation. Configuration options and translation have not been added at this point. Read more about the plugin here https://ranbuch.github.io/accessibility/

To test you will need to
1. apply the patch
2. composer install
3. npm i
4. Discover the plugin
5. Enable the plugin

(I am not a fan of these tools but other people are - so I am offering it for evaluation.)