Remove setFieldAttribute `action` on `catid` in com_content, …
Since the introduction of the categoryedit form field in the edit
models/views of com_content, com_contact, com_newsfeeds and com_banners
(joomla/joomla-cms@d6673d7), there is no `action` attribute on the
`catid` form field and thus no need to set it anymore.

Furthermore cleanups concerning the `$id` in
`ContentModelArticle::getForm()` are performed. As the
`getState('article.id')` method returns the correct article id, no
matter whether in frontend or backend, and seems to be the standard way
of retrieving the article id, there is no need to fetch it from the
`CMSApplication`'s input.