[4.0] Custom fields prefix and suffix
J4 version of #23499

It is very common to need to display a unit of measurement/value with a field. Currently there is no way to do this without creating custom layouts for every field and every unit when needed.

This PR adds two new param to every custom field "Suffix" and "Prefix" which allows you to add a unit of measurement (or any other text) to be rendered after/before the value of the field.

The new field is disabled by default so three is no b/c issue

### To test
Apply the pr and then create a new custom field and you will see the new Prefix Suffix param in the render options.