[4.0] [a11y] Correctly use aria-describedby
For fields where there is a description we are trying to add an aria-describedby with a value of the _DESC language string.

Unfortunately the code was not doing that.

### Testing
Before applying the patch view the generated source of the search field in the article manager. It will display aria-describedby="filter[search]-desc"

After applying the patch it will correctly be
aria-describeby="Search in title and alias. Prefix with ID: or AUTHOR: to search for an article ID or article author."

The display of this text with a tooltip or similar is beyond the scope of this pr