[4.0] [a11y] Remove title from fieldset
The title attribute is generally not very accessible. Keyboard users and touch screen users will never see it.

It is particularly useless to add a title which is the same as other text already present as that can result in a screen reader announcing the same text twice.

This PR removes the redundant title attribute from two fieldsets as the same text is correctly present in the fieldset legend