[4.0] [a11y] Remove redundant title from site components
The title attribute is generally not very accessible. Keyboard users and touch screen users will never see it.

It is particularly useless to add a title which is the same as other text already present as that can result in a screen reader announcing the same text twice.

This PR removes the redundant title attribute where the same text is already displayed and removes/replaces one language string where it wwas not needed

NOTE this is just a first pass at removing the low hanging fruit and a further pr will be needed when the tooltips are reviewed