Fix overlap of editor and options
Hides the header and options in full screen mode to prevent overlap over the editor area in full-screen mode.

Another possible alternative is displaying both the options as well as the editor in full-screen mode but purely a user preference. Because the need for save and other options is not as important in full-screen mode.