[4.0] Password strength [a11y]
When creating a user in the adin we can see a password strength meter.

Currently the meter has an aria-describedby value of the following text.

This is not correct as the meter never receives focus. The text needs to have an aria-live value so that when the text is updated it is announced.

This PR corrects that using the technique from https://www.w3.org/WAI/tutorials/forms/examples/password/

There is no visual change
Viewing the source you will see that the aria-describedby in the progress bar div is now an id
And that the aria-live polite value has been added to the text div