Add the latest status of PR #22446 on top of the 4.0-dev branch, squashed.
This already contains adjustments for Joomla! 4, e.g. changed class files / paths, etc.
This does not yet contain any tests made or bugfixes against the 4.0 branch. This will
be done after this commit and possible bugfixes will be made in standalone commits.
So this basically only contains the code from #22446, adjusted to the 4.0 directory structure.