[4.0] Remove module description popover
When you go to select a NEW module you have a list of all the modules with a very short 90 character truncated description.

If you hover over that description you get a popover with a short 200 character truncated description.

This popover is not accessible - I doubt even most sighted users will know that it even exists as there is no visual indicator that it exists either.

The original reason for the 90 character truncation was back in the day that this was a modal and there was limited space. That is no longer true  but we do need to have some form of truncation as some 3pd write essays for the description ;)

This PR removes the popover and changes the displayed description to the 200 character version