[4.0] com_config options
This is for the options screens for all components. ie the view you get when you click on the options button.
This PR should have no visible impact. It just changes from rendering the input and label as individual items to rendering it as a field which is what we do (almost) everywhere else