[4.0] Smart Search indexer [a11y]
The popup that is loaded when you index the content has a message that is updated during the index process

COM_FINDER_INDEXER_MESSAGE_INIT="The indexer is starting. Do not close this window."
COM_FINDER_INDEXER_MESSAGE_RUNNING="Your content is being indexed. Do not close this window."
COM_FINDER_INDEXER_MESSAGE_OPTIMIZE="The index tables are being optimised for the best possible performance. Do not close this window."
COM_FINDER_INDEXER_MESSAGE_COMPLETE="The indexing process is complete. It is now safe to close this window."

This changing message is invisible to screen readers unless we add an aria-live attribute to the enclosing element. Normally we would have aria-live=polite but in this case there is nothing else on the screen so aria-live=assertive is appropriate