[4.0] Missing layout in notes
This is a bit tricky to test. Create a new note and keep it open until the session dies.

When you log back in you will be redirected to `index.php?option=com_users&view=note&id=0`

This will give a 500 Layout default not found.

Apply the pr and retest. The note will now be opened correctly as expected