[4.0] Searchtools Descriptions (#24892)
In Joomla 3 all the search boxes in the list views in the admin had a descriptive tooltip. This was added by javascript. This tooltip is missing in J4 as we dont have the same js. As a temporary measure the descriptions for the search boxes was not displayed.

This PR puts the descriptive tooltip back WITHOUT using any js and in a fully accessible way. The description is displayed when the field gets focus.

This is fully backwards compatible with the tooltips displaying descriptions in J3 including when there is html in the string.

As the css has changed testing will require npm i

In addition the descriptions rendered on screen for other fields is now escaped properly.