[4.0] Update check alerts (#24893)
In J3 the quickicon for the new version etc where hidden in the bottom left of the control panel so we had a popup alert as well. As the quickicons are now bigger and right at the top we don't need the popup as well.

This PR removes the popup and associated strings and ensures that the icons get the red danger instead of the brown warning and also standardises the 3 strings so that the count is in the same place.