[4.0] modal toolbar layout
As far as I can tell this layout is only used `administrator/index.php?option=com_banners&view=tracks` for the export button.

If you look at the screenshots you will see that the export button is indented more than the new button. This is because there is some css for the first button.

`.subhead .btn-toolbar>:first-child {
    margin-left: 0;
}`

This should be applied to the export button as well but its not hence the indentation.

This is caused by the layout rendering the div for the modal before the button. This means that the button is no longer the `first-child`

This simple PR changes the order of the div and the button to resolve this

NOTE any issues with the rendering of the modal are beyond the scope of this PR