display the master-child associations in the list views
In order to display the Master-Child relationship of the associations in the lists, the layout was adapted.
So now all children of an association are displayed, if the item in the row is a master item. Otherwise, only the current child-item is displayed with its master item, since there is no direct relationship between the children.
In addition, the links of the association items in the Associations View have been adjusted so that the reference is now oriented to the master item and target to the children.