[4.0] Select modal button
When creating many types of menu we have the option to Select, Edit, Create an item

For some reason, perhaps legacy, the title for the Select was "Select or Change" but you can only select. There is no ability to change - to do that you have to clear.

This pr therefore corrects the title on the button and the title on the modal page and then removes the unused strings