[4.0] Render label [a11y]
There is an option "hiddenLabel" for a form field. But instead of "hiding" the label it doesn't render the label at all. This causes accessibility issues because every form field input must have an associated label.

This PR changes it so that instead of the label not being rendered at all it is rendered with a class of sr-only instead.

### Testing
Add the following field to any form
`
		<field
			name="test"
			type="text"
			label="Hidden"
			hiddenLabel="true"
		/>
'
Before and After applying the PR there will be no visual change BUT if you view the source you will see that the code has changed

### Before
`
<div class="control-group">
	<div class="controls">
		<input type="text" name="jform[test]" id="jform_test" value="" class="form-control">
	</div>
</div>
`
### After
`
<div class="control-group">
	<div class="control-label sr-only">
        <label id="jform_test-lbl" for="jform_test">Hidden</label>
	</div>
	<div class="controls has-success">
	<input type="text" name="jform[test]" id="jform_test" value="" class="form-control valid form-control-success" aria-invalid="false">
	</div>
</div>
`